requirements.txt
pandas
scikit-learn
transformers
datasets
evaluate
torch
pdfplumber
PyMuPDF
Pillow
pytesseract
opencv-python
numpy
